Added units tests for you_have_told_us snippet and moved to better location.
