+++ app/uk/gov/hmrc/payetaxcalculatorfrontend/views/quickcalc/quick_calc_form.scala.html
@@ -1,6 +1,6 @@
-@import uk.gov.hmrc.payetaxcalculatorfrontend.views.helpers.YouHaveToldUsItem
+@import uk.gov.hmrc.payetaxcalculatorfrontend.views.quickcalc.YouHaveToldUsItem
 @import uk.gov.hmrc.payetaxcalculatorfrontend.views.html.main_template
-@import uk.gov.hmrc.payetaxcalculatorfrontend.views.html.helpers.you_have_told_us
+@import uk.gov.hmrc.payetaxcalculatorfrontend.views.html.quickcalc.you_have_told_us
 
 
 @()(implicit request: Request[_], messages: Messages)
