+++ app/uk/gov/hmrc/payetaxcalculatorfrontend/views/quickcalc/you_have_told_us.scala.html
@@ -1,4 +1,4 @@
-@import uk.gov.hmrc.payetaxcalculatorfrontend.views.helpers.YouHaveToldUsItem
+@import uk.gov.hmrc.payetaxcalculatorfrontend.views.quickcalc.YouHaveToldUsItem
 
 @(items: List[YouHaveToldUsItem])(implicit request: Request[_], messages: Messages)
